Fix GCC 14 -Warray-bounds and -Wuseless-cast warnings
Md5: restructure block loop so remaining = size % 64, making the
[0,63] range visible to GCC's value-range analysis and preventing a
false-positive out-of-bounds warning on the pad byte write.

RecordParser: drop redundant size_t casts on uint64_t subtraction.

+++ src/tools/zmi-index/RecordParser.hpp
@@ -217,7 +217,7 @@ private:
 
     [[nodiscard]] std::size_t BufferAvailable() const noexcept
     {
-        return std::size(buf_) - static_cast<std::size_t>(cursor_ - windowStart_);
+        return std::size(buf_) - (cursor_ - windowStart_);
     }
 
     [[nodiscard]] const std::byte* CursorPtr() const noexcept
@@ -229,7 +229,7 @@ private:
 
     void Compact()
     {
-        const std::size_t consumed{static_cast<std::size_t>(cursor_ - windowStart_)};
+        const std::size_t consumed{cursor_ - windowStart_};
         if (consumed == 0U) {
             return;
         }
